Organize the code structure
